fix memory leak in MagicHit hook
MagicHit hooks the call to add the effect to the target's active effect list. At this point the new ActiveEffect has already been allocated and initialized, so preventing it from being added results in a leak.

Instead, add the effect as normal, but mark it as dispelled to prevent any damage and allow it to be cleaned up by the next update.